feat (update composite view): make a separate script to update a composite view
